// web/src/generated/models/GetObjectInfoResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GetObjectInfoResponse = {
    bucket?: string;
    chunks?: number;
    description?: string;
    key?: string;
    modified?: string;
    size?: number;
};