// web/src/generated/models/GetObjectResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GetObjectResponse = {
    bucket?: string;
    /**
     * Base64 encoded data
     */
    data?: string;
    key?: string;
    modified?: string;
    size?: number;
};