// web/src/generated/models/ObjectStoreBucket.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ObjectStoreBucket = {
    description?: string;
    max_bytes?: number;
    name?: string;
    num_buckets?: number;
    num_values?: number;
    storage?: string;
};